% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarWards.R
\docType{package}
\name{StarWards}
\alias{StarWards}
\alias{StarWards-package}
\title{May the CSS With You}
\description{
May the CSS With You
}
\seealso{
\href{https://polarnotion.github.io/starwarsintro/}{StarWarsIntro.css}
\href{https://www.flickr.com/photos/internetarchivebookimages/14760845556/}{BackImage from page 218 of "Astronomy for amateurs" (1904)}
}
